# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface AssetPrice {
  symbol: string;
  name: string;
  price: number;
  change24h: number;
  changePercent24h: number;
  volume24h: number;
  marketCap: number;
  high24h: number;
  low24h: number;
  updatedAt: string;
}

export interface Candle {
  /** Unix timestamp in seconds */
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface Account {
  userId: string;
  cashBalance: number;
  totalDeposited: number;
  createdAt: string;
}

export interface Holding {
  symbol: string;
  name: string;
  quantity: number;
  averageCost: number;
  currentPrice: number;
  currentValue: number;
  totalCost: number;
  unrealizedPnl: number;
  unrealizedPnlPercent: number;
}

export interface Portfolio {
  holdings: Holding[];
}

export interface AllocationItem {
  symbol: string;
  name: string;
  percentage: number;
  value: number;
}

export interface PortfolioSummary {
  cashBalance: number;
  totalPortfolioValue: number;
  totalInvested: number;
  totalUnrealizedPnl: number;
  totalUnrealizedPnlPercent: number;
  allTimePnl: number;
  allTimePnlPercent: number;
  dayPnl: number;
  dayPnlPercent: number;
  allocation: AllocationItem[];
}

export type OrderSide = (typeof OrderSide)[keyof typeof OrderSide];

export const OrderSide = {
  buy: "buy",
  sell: "sell",
} as const;

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  market: "market",
  limit: "limit",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  filled: "filled",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: string;
  symbol: string;
  side: OrderSide;
  type: OrderType;
  quantity: number;
  price: number;
  /** @nullable */
  limitPrice?: number | null;
  status: OrderStatus;
  /** @nullable */
  filledAt: string | null;
  createdAt: string;
}

export type PlaceOrderBodySide =
  (typeof PlaceOrderBodySide)[keyof typeof PlaceOrderBodySide];

export const PlaceOrderBodySide = {
  buy: "buy",
  sell: "sell",
} as const;

export type PlaceOrderBodyType =
  (typeof PlaceOrderBodyType)[keyof typeof PlaceOrderBodyType];

export const PlaceOrderBodyType = {
  market: "market",
  limit: "limit",
} as const;

export interface PlaceOrderBody {
  symbol: string;
  side: PlaceOrderBodySide;
  type: PlaceOrderBodyType;
  /** @minimum 0 */
  quantity: number;
  /** @nullable */
  limitPrice?: number | null;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type GetCandleDataParams = {
  interval?: GetCandleDataInterval;
  limit?: number;
};

export type GetCandleDataInterval =
  (typeof GetCandleDataInterval)[keyof typeof GetCandleDataInterval];

export const GetCandleDataInterval = {
  "1m": "1m",
  "5m": "5m",
  "15m": "15m",
  "1h": "1h",
  "4h": "4h",
  "1d": "1d",
} as const;

export type GetOrdersParams = {
  symbol?: string;
  status?: GetOrdersStatus;
  limit?: number;
};

export type GetOrdersStatus =
  (typeof GetOrdersStatus)[keyof typeof GetOrdersStatus];

export const GetOrdersStatus = {
  pending: "pending",
  filled: "filled",
  cancelled: "cancelled",
  all: "all",
} as const;