# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce.string().optional(),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

/**
 * @summary Exchange a mobile OIDC code for a session token
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().url().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — Bearer <sid>."),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current prices for all tracked assets
 */
export const GetPricesResponseItem = zod.object({
  symbol: zod.string(),
  name: zod.string(),
  price: zod.number(),
  change24h: zod.number(),
  changePercent24h: zod.number(),
  volume24h: zod.number(),
  marketCap: zod.number(),
  high24h: zod.number(),
  low24h: zod.number(),
  updatedAt: zod.coerce.date(),
});
export const GetPricesResponse = zod.array(GetPricesResponseItem);

/**
 * @summary Get candlestick OHLCV data for a symbol
 */
export const GetCandleDataParams = zod.object({
  symbol: zod.coerce.string(),
});

export const getCandleDataQueryIntervalDefault = `1h`;
export const getCandleDataQueryLimitDefault = 100;

export const GetCandleDataQueryParams = zod.object({
  interval: zod
    .enum(["1m", "5m", "15m", "1h", "4h", "1d"])
    .default(getCandleDataQueryIntervalDefault),
  limit: zod.coerce.number().default(getCandleDataQueryLimitDefault),
});

export const GetCandleDataResponseItem = zod.object({
  time: zod.number().describe("Unix timestamp in seconds"),
  open: zod.number(),
  high: zod.number(),
  low: zod.number(),
  close: zod.number(),
  volume: zod.number(),
});
export const GetCandleDataResponse = zod.array(GetCandleDataResponseItem);

/**
 * @summary Get the authenticated user's account info and balance
 */
export const GetAccountResponse = zod.object({
  userId: zod.string(),
  cashBalance: zod.number(),
  totalDeposited: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get the authenticated user's portfolio holdings
 */
export const GetPortfolioResponse = zod.object({
  holdings: zod.array(
    zod.object({
      symbol: zod.string(),
      name: zod.string(),
      quantity: zod.number(),
      averageCost: zod.number(),
      currentPrice: zod.number(),
      currentValue: zod.number(),
      totalCost: zod.number(),
      unrealizedPnl: zod.number(),
      unrealizedPnlPercent: zod.number(),
    }),
  ),
});

/**
 * @summary Get portfolio summary with total value, P&L and allocation
 */
export const GetPortfolioSummaryResponse = zod.object({
  cashBalance: zod.number(),
  totalPortfolioValue: zod.number(),
  totalInvested: zod.number(),
  totalUnrealizedPnl: zod.number(),
  totalUnrealizedPnlPercent: zod.number(),
  allTimePnl: zod.number(),
  allTimePnlPercent: zod.number(),
  dayPnl: zod.number(),
  dayPnlPercent: zod.number(),
  allocation: zod.array(
    zod.object({
      symbol: zod.string(),
      name: zod.string(),
      percentage: zod.number(),
      value: zod.number(),
    }),
  ),
});

/**
 * @summary Get the authenticated user's order history
 */
export const getOrdersQueryLimitDefault = 50;

export const GetOrdersQueryParams = zod.object({
  symbol: zod.coerce.string().optional(),
  status: zod.enum(["pending", "filled", "cancelled", "all"]).optional(),
  limit: zod.coerce.number().default(getOrdersQueryLimitDefault),
});

export const GetOrdersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  symbol: zod.string(),
  side: zod.enum(["buy", "sell"]),
  type: zod.enum(["market", "limit"]),
  quantity: zod.number(),
  price: zod.number(),
  limitPrice: zod.number().nullish(),
  status: zod.enum(["pending", "filled", "cancelled"]),
  filledAt: zod.coerce.date().nullable(),
  createdAt: zod.coerce.date(),
});
export const GetOrdersResponse = zod.array(GetOrdersResponseItem);

/**
 * @summary Place a market or limit order
 */
export const placeOrderBodyQuantityMin = 0;

export const PlaceOrderBody = zod.object({
  symbol: zod.string(),
  side: zod.enum(["buy", "sell"]),
  type: zod.enum(["market", "limit"]),
  quantity: zod.number().min(placeOrderBodyQuantityMin),
  limitPrice: zod.number().nullish(),
});

/**
 * @summary Cancel a pending limit order
 */
export const CancelOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const CancelOrderResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  symbol: zod.string(),
  side: zod.enum(["buy", "sell"]),
  type: zod.enum(["market", "limit"]),
  quantity: zod.number(),
  price: zod.number(),
  limitPrice: zod.number().nullish(),
  status: zod.enum(["pending", "filled", "cancelled"]),
  filledAt: zod.coerce.date().nullable(),
  createdAt: zod.coerce.date(),
});